print_list

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -1,17 +1,15 @@
 #include "lists.h"
-#include <stdlib.h>
-#include <string.h>
-#include <stdio.h>
 
 /**
  * print_list - prints all the elements of a list_t list
- * @h: the input list
- * Return: number of nodes.
+ * @h: pointer to the head of the list
+ *
+ * Return: the number of nodes in the list
  */
+
 size_t print_list(const list_t *h)
 {
-	int count;
-
+	size_t count;
 
 	count = 0;
 
@@ -19,14 +17,10 @@ size_t print_list(const list_t *h)
 	{
 		if (h->str == NULL)
 			printf("[0] (nil)\n");
-
 		else
-			printf("[%d] %s\n", h->len, h->str);
-
+			printf("[%u] %s\n", h->len, h->str);
 		count++;
-
 		h = h->next;
 	}
-
 	return (count);
 }
